+++ src/main/java/com/garbagemule/MobArena/ArenaMasterImpl.java
@@ -532,6 +532,8 @@ public class ArenaMasterImpl implements ArenaMaster
         if (arenaNames == null || arenaNames.isEmpty()) {
             return;
         }
+
+        List<Arena> arenas = new ArrayList<>();
         for (String arenaName : arenaNames) {
             Arena arena = getArenaWithName(arenaName);
             if (arena != null) continue;
@@ -539,7 +541,31 @@ public class ArenaMasterImpl implements ArenaMaster
             String arenaWorld = config.getString("arenas." + arenaName + ".settings.world", "");
             if (!arenaWorld.equals(worldName)) continue;
 
-            loadArena(arenaName);
+            Arena loaded = loadArena(arenaName);
+            if (loaded != null) {
+                arenas.add(loaded);
+            }
+        }
+
+        reportOverlappingRegions(arenas);
+    }
+
+    private void reportOverlappingRegions(List<Arena> arenas) {
+        // If we iterate the upper/lower triangular matrix of the cartesian
+        // product of the arena list, we avoid not only duplicate reports like
+        // "a vs. b and b vs. a", but also "self comparisons" (j = i + 1).
+        for (int i = 0; i < arenas.size(); i++) {
+            Arena a = arenas.get(i);
+            for (int j = i + 1; j < arenas.size(); j++) {
+                Arena b = arenas.get(j);
+                if (a.getRegion().intersects(b.getRegion())) {
+                    plugin.getLogger().warning(String.format(
+                        "Regions of arenas '%s' and '%s' overlap!",
+                        a.configName(),
+                        b.configName()
+                    ));
+                }
+            }
         }
     }